+++ src/main/java/br/com/fiap/soat/techChallenge/api/ClienteController.java
@@ -1,6 +1,6 @@
 package br.com.fiap.soat.techChallenge.api;
 
-import br.com.fiap.soat.techChallenge.api.model.response.ClienteResponse;
+import br.com.fiap.soat.techChallenge.responses.ClienteResponse;
 import br.com.fiap.soat.techChallenge.requests.CadastrarClienteRequest;
 import br.com.fiap.soat.techChallenge.entities.Cliente;
 import br.com.fiap.soat.techChallenge.exceptions.ClienteNaoEncontradoException;
