+++ src/main/java/br/com/fiap/soat/techChallenge/api/PedidoController.java
@@ -1,7 +1,7 @@
 package br.com.fiap.soat.techChallenge.api;
 
 import br.com.fiap.soat.techChallenge.requests.PedidoRequest;
-import br.com.fiap.soat.techChallenge.api.model.response.PedidoResponse;
+import br.com.fiap.soat.techChallenge.responses.PedidoResponse;
 import br.com.fiap.soat.techChallenge.entities.Pedido;
 import br.com.fiap.soat.techChallenge.interfaces.usecases.FazerPedidoUseCasePort;
 import br.com.fiap.soat.techChallenge.interfaces.usecases.ObterPedidosUseCasePort;
