+++ src/main/java/br/com/fiap/soat/techChallenge/responses/ProdutoResponse.java
@@ -1,4 +1,4 @@
-package br.com.fiap.soat.techChallenge.api.model.response;
+package br.com.fiap.soat.techChallenge.responses;
 import br.com.fiap.soat.techChallenge.entities.Produto;
 
 import java.util.UUID;
